requirements.txt
# Dependências principais
streamlit==1.26.0
streamlit-chatbox==0.1.1
streamlit-elements==0.1.0
streamlit-apexcharts==0.0.3

# Componentes de IA e ML
tensorflow==2.13.0
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.25.2
matplotlib==3.7.2
nltk==3.8.1
transformers==4.33.0
huggingface-hub==0.17.1

# Comunicação e API
fastapi==0.103.1
uvicorn==0.23.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.3.0

# Banco de dados
pymongo==4.5.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
redis==4.6.0

# Blockchain e Web3
web3==6.9.0
eth-typing==3.4.0
eth-utils==2.2.0

# Segurança e autenticação
pyjwt==2.8.0
cryptography==41.0.3

# Testes e qualidade de código
pytest==7.4.0
flake8==6.1.0
pytest-cov==4.1.0

# Utilitários
python-dateutil==2.8.2
tqdm==4.66.1 